fix: reset the thread sequence when loading the database
When the database was loaded the `max_thread_id` is always set to `1`. When the
database is loaded after a restart of app, this was causing the thread id
sequence to get reset and overwrite keys in the database.

Now when the database is loaded we are getting the largest key and storing that
as the sequence number so when we get the next ID that number is incremented and
there are no collisions.