fix: get email from other folder other than the first one
When getting the content of an email, we are trying to look in every folder in
the account to try and find the email key we want the content for. However, if
the key is not found in the fist folder then the function returns with an error.

Now we are skipping all the errors that are generated from a not found key and
then returning a not found error at the end after all the folders in an account
have been checked.